# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SupportFlow API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/threads": {
      "get": {
        "summary": "Threads",
        "operationId": "threads_threads_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreadView"
                  },
                  "type": "array",
                  "title": "Response Threads Threads Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "New Thread",
        "operationId": "new_thread_threads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/threads/{thread_id}": {
      "get": {
        "summary": "Thread",
        "operationId": "thread_threads__thread_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "get": {
        "summary": "Documents",
        "operationId": "documents_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "type": "array",
                  "title": "Response Documents Documents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/feedback": {
      "post": {
        "summary": "Feedback",
        "operationId": "feedback_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tickets": {
      "post": {
        "summary": "Tickets",
        "operationId": "tickets_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolEvent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/evals/run": {
      "post": {
        "summary": "Evals",
        "operationId": "evals_evals_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/monitoring": {
      "get": {
        "summary": "Monitoring",
        "operationId": "monitoring_monitoring_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Ack": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "Ack"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "filters": {
            "$ref": "#/components/schemas/Filters"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "response_id": {
            "type": "string",
            "title": "Response Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "tool_events": {
            "items": {
              "$ref": "#/components/schemas/ToolEvent"
            },
            "type": "array",
            "title": "Tool Events"
          },
          "needs_escalation": {
            "type": "boolean",
            "title": "Needs Escalation"
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          },
          "route": {
            "type": "string",
            "title": "Route"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "trajectory": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trajectory"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id",
          "response_id",
          "answer",
          "citations",
          "tool_events",
          "needs_escalation",
          "route",
          "trace_id",
          "request_id",
          "latency_ms",
          "trajectory"
        ],
        "title": "ChatResponse"
      },
      "Citation": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "doc_id",
          "title",
          "product",
          "version",
          "source_type",
          "trust_level",
          "chunk_id",
          "content",
          "score",
          "source_path"
        ],
        "title": "Citation"
      },
      "DocumentCreate": {
        "properties": {
          "markdown": {
            "type": "string",
            "maxLength": 60000,
            "minLength": 20,
            "title": "Markdown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "DocumentCreate"
      },
      "DocumentInfo": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "product": {
            "type": "string",
            "title": "Product"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level"
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "doc_id",
          "title",
          "product",
          "version",
          "source_type",
          "trust_level",
          "source_path"
        ],
        "title": "DocumentInfo"
      },
      "EvalRequest": {
        "properties": {
          "suite": {
            "type": "string",
            "const": "structural",
            "title": "Suite",
            "default": "structural"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EvalRequest"
      },
      "EvalResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "report"
        ],
        "title": "EvalResponse"
      },
      "FeedbackRequest": {
        "properties": {
          "response_id": {
            "type": "string",
            "title": "Response Id"
          },
          "helpful": {
            "type": "boolean",
            "title": "Helpful"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "response_id",
          "helpful"
        ],
        "title": "FeedbackRequest"
      },
      "Filters": {
        "properties": {
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "trust_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "official",
                  "internal",
                  "untrusted"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Level",
            "default": "official"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Filters"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "role",
          "content"
        ],
        "title": "MessageView"
      },
      "MonitoringResponse": {
        "properties": {
          "requests": {
            "type": "integer",
            "title": "Requests"
          },
          "mean_latency_ms": {
            "type": "number",
            "title": "Mean Latency Ms"
          },
          "escalation_rate": {
            "type": "number",
            "title": "Escalation Rate"
          },
          "error_rate": {
            "type": "number",
            "title": "Error Rate"
          },
          "recent": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requests",
          "mean_latency_ms",
          "escalation_rate",
          "error_rate",
          "recent"
        ],
        "title": "MonitoringResponse"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dependencies": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Dependencies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "dependencies"
        ],
        "title": "StatusResponse"
      },
      "ThreadCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title",
            "default": "New conversation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ThreadCreate"
      },
      "ThreadView": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageView"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id",
          "title"
        ],
        "title": "ThreadView"
      },
      "TicketRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "summary": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 3,
            "title": "Summary"
          },
          "category": {
            "type": "string",
            "enum": [
              "security",
              "privacy",
              "billing",
              "technical"
            ],
            "title": "Category",
            "default": "technical"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id",
          "summary"
        ],
        "title": "TicketRequest"
      },
      "ToolError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ToolError"
      },
      "ToolEvent": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "data": {
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolError"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "ok"
        ],
        "title": "ToolEvent"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}